Add Deck type with single draw function
Deck inherits from List because it's pretty much a list with some helper
functions, or at least, that's the plan for now. I added an IDeck
interface type for consumption, that way we can swap out the
implementation of Deck down the line. It would probably be better to
just have Deck implement pass-throughs to the methods we need, and hide
everything else via an internal base class or something. idk.